+++ src/main/java/roomescape/controller/TimeController.java
@@ -1,5 +1,7 @@
 package roomescape.controller;
 
+import java.net.URI;
+import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,31 +12,28 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.time.TimeRequest;
 import roomescape.dto.time.TimeResponse;
-import roomescape.service.ReservationTimeService;
-
-import java.net.URI;
-import java.util.List;
+import roomescape.service.TimeService;
 
 @RestController
 @RequestMapping("/times")
 public class TimeController {
 
-    private final ReservationTimeService reservationTimeService;
+    private final TimeService timeService;
 
-    public TimeController(ReservationTimeService reservationTimeService) {
-        this.reservationTimeService = reservationTimeService;
+    public TimeController(TimeService timeService) {
+        this.timeService = timeService;
     }
 
     @GetMapping
     public ResponseEntity<List<TimeResponse>> readTimes() {
-        List<TimeResponse> timeResponses = reservationTimeService.findAllTimes();
+        List<TimeResponse> timeResponses = timeService.findAllTimes();
 
         return ResponseEntity.ok(timeResponses);
     }
 
     @PostMapping
     public ResponseEntity<TimeResponse> createTime(@RequestBody TimeRequest timeRequest) {
-        TimeResponse timeResponse = reservationTimeService.createTime(timeRequest);
+        TimeResponse timeResponse = timeService.createTime(timeRequest);
 
         return ResponseEntity.created(URI.create("/times/" + timeResponse.id()))
                 .body(timeResponse);
@@ -42,7 +41,7 @@ public class TimeController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteTime(@PathVariable Long id) {
-        reservationTimeService.deleteTime(id);
+        timeService.deleteTime(id);
 
         return ResponseEntity.noContent().build();
     }
