+++ src/main/java/roomescape/service/ReservationService.java
@@ -6,18 +6,16 @@ import roomescape.domain.reservation.Reservation;
 import roomescape.domain.time.Time;
 import roomescape.dto.reservation.ReservationRequest;
 import roomescape.dto.reservation.ReservationResponse;
-import roomescape.dto.time.TimeRequest;
-import roomescape.dto.time.TimeResponse;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.TimeRepository;
 
 @Service
-public class ReservationTimeService {
+public class ReservationService {
 
     private final ReservationRepository reservationRepository;
     private final TimeRepository timeRepository;
 
-    public ReservationTimeService(ReservationRepository reservationRepository, TimeRepository timeRepository) {
+    public ReservationService(ReservationRepository reservationRepository, TimeRepository timeRepository) {
         this.reservationRepository = reservationRepository;
         this.timeRepository = timeRepository;
     }
@@ -40,23 +38,5 @@ public class ReservationTimeService {
 
     public void deleteReservation(Long id) {
         reservationRepository.delete(id);
-    }
-
-    public List<TimeResponse> findAllTimes() {
-        return timeRepository.findAll()
-                .stream()
-                .map(TimeResponse::from)
-                .toList();
-    }
-
-    public TimeResponse createTime(TimeRequest timeRequest) {
-        Time time = timeRequest.toTime();
-        Time savedTime = timeRepository.save(time);
-
-        return TimeResponse.from(savedTime);
-    }
-
-    public void deleteTime(Long id) {
-        timeRepository.delete(id);
     }
 }
