+++ src/main/java/roomescape/service/TimeService.java
@@ -1,0 +1,36 @@
+package roomescape.service;
+
+import java.util.List;
+import org.springframework.stereotype.Service;
+import roomescape.domain.time.Time;
+import roomescape.dto.time.TimeRequest;
+import roomescape.dto.time.TimeResponse;
+import roomescape.repository.TimeRepository;
+
+@Service
+public class TimeService {
+
+    private final TimeRepository timeRepository;
+
+    public TimeService(TimeRepository timeRepository) {
+        this.timeRepository = timeRepository;
+    }
+
+    public List<TimeResponse> findAllTimes() {
+        return timeRepository.findAll()
+                .stream()
+                .map(TimeResponse::from)
+                .toList();
+    }
+
+    public TimeResponse createTime(TimeRequest timeRequest) {
+        Time time = timeRequest.toTime();
+        Time savedTime = timeRepository.save(time);
+
+        return TimeResponse.from(savedTime);
+    }
+
+    public void deleteTime(Long id) {
+        timeRepository.delete(id);
+    }
+}
